Update schema-accounts to schema
