--- packages/schema-accounts/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@cord.network/schema-accounts",
-  "version": "0.9.3-1rc4",
-  "description": "Accounts Based Schema Repository Management",
-  "main": "./lib/cjs/index.js",
-  "module": "./lib/esm/index.js",
-  "types": "./lib/cjs/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./lib/esm/index.js",
-      "require": "./lib/cjs/index.js"
-    }
-  },
-  "files": [
-    "lib/**/*"
-  ],
-  "scripts": {
-    "clean": "rimraf ./lib",
-    "build": "yarn clean && yarn build:ts",
-    "build:ts": "yarn build:cjs && yarn build:esm",
-    "build:cjs": "tsc --declaration -p tsconfig.build.json && echo '{\"type\":\"commonjs\"}' > ./lib/cjs/package.json",
-    "build:esm": "tsc --declaration -p tsconfig.esm.json && echo '{\"type\":\"module\"}' > ./lib/esm/package.json"
-  },
-  "repository": "github:dhiway/cord.js",
-  "engines": {
-    "node": ">=20.0"
-  },
-  "author": "",
-  "bugs": "https://github.com/dhiway/cord.js/issues",
-  "homepage": "https://github.com/dhiway/cord.js#readme",
-  "devDependencies": {
-    "@types/uuid": "^9.0.8",
-    "rimraf": "^5.0.5",
-    "testcontainers": "^10.7.1",
-    "typescript": "^5.3.3"
-  },
-  "dependencies": {
-    "@cord.network/augment-api": "workspace:*",
-    "@cord.network/config": "workspace:*",
-    "@cord.network/did": "workspace:*",
-    "@cord.network/identifier": "workspace:*",
-    "@cord.network/network": "workspace:*",
-    "@cord.network/types": "workspace:*",
-    "@cord.network/utils": "workspace:*"
-  }
-}
