--- packages/schema-accounts/src/index.ts
@@ -1,3 +1,0 @@
-export * from './SchemaAccounts.js'
-export * from './SchemaAccounts.chain.js'
-export * as TypeSchemaAccounts from './SchemaAccounts.types.js'
