--- packages/schema-accounts/tsconfig.build.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-
-  "compilerOptions": {
-    "module": "CommonJS",
-    "outDir": "./lib/cjs"
-  },
-
-  "include": [
-    "src/**/*.ts", "src/**/*.js"
-  ],
-
-  "exclude": [
-    "coverage",
-    "**/*.spec.ts",
-  ]
-}
