--- packages/schema-accounts/tsconfig.esm.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.build.json",
-  "compilerOptions": {
-    "module": "ES6",
-    "outDir": "./lib/esm"
-  }
-}
