+++ packages/schema/tsconfig.build.json
@@ -11,18 +11,7 @@
   ],
 
   "exclude": [
-    "node_modules",
     "coverage",
-    "build",
-    "scripts",
-    "acceptance-tests",
-    "webpack",
-    "jest",
-    "src/setupTests.ts",
-    "jest.config.js",
-    "docs",
     "**/*.spec.ts",
-    "src/__integrationtests__",
-    "**/test"
   ]
 }
